refactor: extract shared control-flow graph into internal/utils/cfg
The graph construction built for no-useless-assignment was generic apart
from the two calls that recorded variable reads and writes. Move it into
internal/utils/cfg, parameterized over a consumer-supplied event type, and
let consumers plug in through Hooks[E]{Read, Write}.

Builder.Emit returns the block and index it wrote to, so a consumer can
record where an event landed - no-useless-assignment needs this because
the doubled finally layout gives one assignment two sites. Block.Index
lets a consumer keep per-block analysis state in a slice rather than on
the block itself.

no-useless-assignment keeps only its events and the backward liveness
dataflow, now in liveness.go.
